:recycle: refactor : rename package name to singular
