+++ src/main/java/org/runimo/runimo/auth/controller/AuthController.java
@@ -15,11 +15,11 @@ import org.runimo.runimo.auth.controller.request.KakaoLoginRequest;
 import org.runimo.runimo.auth.service.OidcService;
 import org.runimo.runimo.auth.service.SignUpUsecase;
 import org.runimo.runimo.auth.service.TokenRefreshService;
-import org.runimo.runimo.auth.service.dtos.AuthResponse;
-import org.runimo.runimo.auth.service.dtos.AuthResult;
-import org.runimo.runimo.auth.service.dtos.AuthStatus;
-import org.runimo.runimo.auth.service.dtos.SignupUserResponse;
-import org.runimo.runimo.auth.service.dtos.TokenPair;
+import org.runimo.runimo.auth.service.dto.AuthResponse;
+import org.runimo.runimo.auth.service.dto.AuthResult;
+import org.runimo.runimo.auth.service.dto.AuthStatus;
+import org.runimo.runimo.auth.service.dto.SignupUserResponse;
+import org.runimo.runimo.auth.service.dto.TokenPair;
 import org.runimo.runimo.common.response.ErrorResponse;
 import org.runimo.runimo.common.response.Response;
 import org.runimo.runimo.common.response.SuccessResponse;
