+++ src/main/java/org/runimo/runimo/auth/service/OidcService.java
@@ -2,7 +2,7 @@ package org.runimo.runimo.auth.service;
 
 import lombok.RequiredArgsConstructor;
 import org.runimo.runimo.auth.service.apple.AppleLoginHandler;
-import org.runimo.runimo.auth.service.dtos.AuthResult;
+import org.runimo.runimo.auth.service.dto.AuthResult;
 import org.runimo.runimo.auth.service.kakao.KakaoLoginHandler;
 import org.springframework.stereotype.Service;
