+++ src/main/java/org/runimo/runimo/auth/service/SignUpUsecase.java
@@ -1,7 +1,7 @@
 package org.runimo.runimo.auth.service;
 
-import org.runimo.runimo.auth.service.dtos.SignupUserResponse;
-import org.runimo.runimo.auth.service.dtos.UserSignupCommand;
+import org.runimo.runimo.auth.service.dto.SignupUserResponse;
+import org.runimo.runimo.auth.service.dto.UserSignupCommand;
 
 public interface SignUpUsecase {
