+++ src/main/java/org/runimo/runimo/auth/service/SignUpUsecaseImpl.java
@@ -7,8 +7,8 @@ import org.runimo.runimo.auth.jwt.JwtResolver;
 import org.runimo.runimo.auth.jwt.JwtTokenFactory;
 import org.runimo.runimo.auth.jwt.SignupTokenPayload;
 import org.runimo.runimo.auth.repository.SignupTokenRepository;
-import org.runimo.runimo.auth.service.dtos.SignupUserResponse;
-import org.runimo.runimo.auth.service.dtos.UserSignupCommand;
+import org.runimo.runimo.auth.service.dto.SignupUserResponse;
+import org.runimo.runimo.auth.service.dto.UserSignupCommand;
 import org.runimo.runimo.user.domain.AppleUserToken;
 import org.runimo.runimo.user.domain.SocialProvider;
 import org.runimo.runimo.user.domain.User;
