+++ src/main/java/org/runimo/runimo/auth/service/TokenRefreshService.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.runimo.runimo.auth.exceptions.UserJwtException;
 import org.runimo.runimo.auth.jwt.JwtResolver;
 import org.runimo.runimo.auth.jwt.JwtTokenFactory;
-import org.runimo.runimo.auth.service.dtos.TokenPair;
+import org.runimo.runimo.auth.service.dto.TokenPair;
 import org.runimo.runimo.common.cache.InMemoryCache;
 import org.runimo.runimo.user.enums.UserHttpResponseCode;
 import org.springframework.beans.factory.annotation.Value;
