+++ src/main/java/org/runimo/runimo/auth/service/dto/AuthStatus.java
@@ -1,4 +1,4 @@
-package org.runimo.runimo.auth.service.dtos;
+package org.runimo.runimo.auth.service.dto;
 
 public enum AuthStatus {
     LOGIN_SUCCESS,
