+++ src/main/java/org/runimo/runimo/auth/service/dto/TokenPair.java
@@ -1,4 +1,4 @@
-package org.runimo.runimo.auth.service.dtos;
+package org.runimo.runimo.auth.service.dto;
 
 public record TokenPair(String accessToken, String refreshToken) {
