+++ src/main/java/org/runimo/runimo/auth/service/kakao/KakaoLoginHandler.java
@@ -11,9 +11,9 @@ import org.runimo.runimo.auth.exceptions.UserJwtException;
 import org.runimo.runimo.auth.jwt.JwtTokenFactory;
 import org.runimo.runimo.auth.repository.SignupTokenRepository;
 import org.runimo.runimo.auth.service.apple.KakaoUserInfo;
-import org.runimo.runimo.auth.service.dtos.AuthResult;
-import org.runimo.runimo.auth.service.dtos.AuthStatus;
-import org.runimo.runimo.auth.service.dtos.TokenPair;
+import org.runimo.runimo.auth.service.dto.AuthResult;
+import org.runimo.runimo.auth.service.dto.AuthStatus;
+import org.runimo.runimo.auth.service.dto.TokenPair;
 import org.runimo.runimo.user.domain.OAuthInfo;
 import org.runimo.runimo.user.domain.SocialProvider;
 import org.runimo.runimo.user.enums.UserHttpResponseCode;
