+++ src/test/java/org/runimo/runimo/auth/controller/AuthControllerTest.java
@@ -15,9 +15,9 @@ import org.runimo.runimo.auth.jwt.JwtResolver;
 import org.runimo.runimo.auth.service.OidcService;
 import org.runimo.runimo.auth.service.SignUpUsecase;
 import org.runimo.runimo.auth.service.TokenRefreshService;
-import org.runimo.runimo.auth.service.dtos.AuthResult;
-import org.runimo.runimo.auth.service.dtos.AuthStatus;
-import org.runimo.runimo.auth.service.dtos.TokenPair;
+import org.runimo.runimo.auth.service.dto.AuthResult;
+import org.runimo.runimo.auth.service.dto.AuthStatus;
+import org.runimo.runimo.auth.service.dto.TokenPair;
 import org.runimo.runimo.configs.ControllerTest;
 import org.runimo.runimo.user.UserFixtures;
 import org.runimo.runimo.user.enums.UserHttpResponseCode;
