+++ src/test/java/org/runimo/runimo/user/api/UserItemAcceptanceTest.java
@@ -23,8 +23,8 @@ import org.runimo.runimo.CleanUpUtil;
 import org.runimo.runimo.auth.controller.request.AuthSignupRequest;
 import org.runimo.runimo.auth.jwt.JwtTokenFactory;
 import org.runimo.runimo.auth.service.SignUpUsecaseImpl;
-import org.runimo.runimo.auth.service.dtos.SignupUserResponse;
-import org.runimo.runimo.auth.service.dtos.TokenPair;
+import org.runimo.runimo.auth.service.dto.SignupUserResponse;
+import org.runimo.runimo.auth.service.dto.TokenPair;
 import org.runimo.runimo.user.controller.request.UseItemRequest;
 import org.runimo.runimo.user.domain.Gender;
 import org.runimo.runimo.user.domain.SocialProvider;
